abbreviate unsigned long literal

+++ cmd/lefty/ws/x11/libfilereq/SelFile.c
@@ -715,7 +715,7 @@ XsraSelFile (
     XtAddGrab (selFile, True, True);
 
     SFdirModTimerId = XtAppAddTimeOut (
-        SFapp, (unsigned long) 1000, SFdirModTimer, (XtPointer) NULL
+        SFapp, 1000ul, SFdirModTimer, (XtPointer) NULL
     );
 
     while (1) {
